Moved most of the code out of main.js
