+++ scripts/components/locations.js
@@ -1,11 +1,11 @@
 import { emptyElement } from "../utilities/general";
-import { locationList } from "../main";
+// import { locationList } from "../main";
 
 let locationContainerEl;
 
 export function init() {
   console.log("Locations init");
-  console.log(locationContainerEl);
+  // console.log(locationContainerEl);
   locationContainerEl = document.getElementById("location-container");
 
   locationContainerEl.addEventListener("click", (event) => {
@@ -47,4 +47,27 @@ function highlightLocationEntry(index, state) {
       }, 1);
     });
   }
+}
+
+function locationItem(itemData, index) {
+  let item = document.createElement("a");
+  item.classList.add("panel-block", "is-active");
+  item.setAttribute("data-lon", itemData.lon);
+  item.setAttribute("data-lat", itemData.lat);
+  item.dataset.locationIndex = index;
+  let leadingIcon = document.createElement("span");
+  leadingIcon.classList.add("panel-icon");
+  let leadingIconImg = document.createElement("i");
+  leadingIconImg.classList.add("fa-sharp", "fa-solid", "fa-location-dot");
+  let itemName = document.createElement("span");
+  itemName.textContent = itemData.name;
+  let deleteIcon = document.createElement("span");
+  deleteIcon.classList.add("panel-icon", "delete");
+  // let deleteIconImg = document.createElement("i");
+  // deleteIconImg.classList.add("fa-sharp", "fa-solid", "fa-xmark");
+
+  leadingIcon.append(leadingIconImg);
+  // deleteIcon.append(deleteIconImg);
+  item.append(leadingIcon, itemName, deleteIcon);
+  return item;
 }
